requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bz2file==0.98
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.1
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
imageio==2.5.0
imgaug==0.2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
MedPy==0.3.0
mistune==0.8.4
mock==3.0.5
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
nibabel==2.4.1
notebook==5.7.6
numpy==1.16.4
opencv-python==4.0.0.21
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pydicom==1.2.2
Pygments==2.3.1
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.3.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==6.0.1
tqdm==4.32.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
